architecture reworked
