+++ registration/src/components/WelcomePage/WelcomePage.jsx
